# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GPSDEV API",
    "description": "Описание взаимодействия с API для формирования отчетов о перемещениях сотрудников и посещениях ими подопечных.",
    "version": "1.0."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api",
      "description": "Development server"
    },
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Development server with gunicorn"
    }
  ],
  "paths": {
...
    "/objects": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Только активные подопечные"
          }
        ],
        "tags": [
          "Objects"
        ],
        "summary": "Список подопечных",
        "description": "Список всех доступных пользователю подопечных. Если добавлен параметр `active=true`, то возвращается список абсолютно всех подопечных, у кого параметр active = true (в том числе из недоступных учреждений), но с ограниченной информацией. Это реализовано для возможности добавлять к себе в таблицу подопечных на обслуживание из других подразделений (например, один ванщик на два подразделения). Подробнее - см. схему ниже.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/Object"
                      },
                      {
                        "$ref": "#/components/schemas/ActiveObject"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Objects"
        ],
        "summary": "Добавить нового подопечного",
        "description": "Добавить нового подопечного",
        "requestBody": {
          "description": "## Добавить нового подопечного\n| Параметр | Описание |\n| -------- | -------- |\n| `name`   | ФИО подопечного |\n| `division` | Номер подразделения, к которому относится подопечный |\n| `latitude`, `longitude` | Координаты нахождения дома подопечного |\n| `address` | Адрес |\n| `no_payments` | Если подопечный обслуживается на частично платной основе, но не доплачивает |\n| `active` | Отображать ли подопечного в списке при добавлении в таблицу в выходов |\n| `phone` | Строка до 200 символов, для хранения любой контактной информации |\n",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{object_id}": {
      "parameters": [
        {
          "in": "path",
          "name": "object_id",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "description": "Object's name_id"
        }
      ],
      "get": {
        "tags": [
          "Objects"
        ],
        "summary": "Получить подопечного",
        "description": "Получение подопечного по object_id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Objects"
        ],
        "summary": "Изменение подопечного",
        "description": "## Изменение свойств подопечного\n| Параметр | Описание |\n| -------- | -------- |\n| `name`   | ФИО подопечного |\n| `division` | Номер подразделения, к которому относится подопечный |\n| `latitude`, `longitude` | Координаты нахождения дома подопечного |\n| `address` | Адрес |\n| `no_payments` | Если подопечный обслуживается на частично платной основе, но не доплачивает |\n| `active` | Отображать ли подопечного в списке при добавлении в таблицу в выходов |\n| `phone` | Строка до 200 символов, для хранения любой контактной информации |\n",
        "requestBody": {
          "description": "A JSON User object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Object"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Objects"
        ],
        "summary": "Удаление подопечного",
        "description": "Удалить подопечного по name_id",
        "responses": {
          "204": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAllowedError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "422"
          },
          "title": {
            "type": "string",
            "example": "validation error"
          },
          "details": {
            "type": "object",
            "example": {}
          }
        }
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "404"
          },
          "title": {
            "type": "string",
            "example": "not found"
          },
          "details": {
            "type": "object",
            "example": {}
          }
        }
      },
      "NotAllowedError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "403"
          },
          "title": {
            "type": "string",
            "example": "not allowed"
          },
          "details": {
            "type": "object",
            "example": {}
          }
        }
      },
...
      "Object": {
        "type": "object",
        "properties": {
          "object_id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "Иванов Иван Петрович"
          },
          "division": {
            "type": "integer",
            "example": 7
          },
          "latitude": {
            "type": "number",
            "example": 55.846497
          },
          "longitude": {
            "type": "number",
            "example": 37.448375
          },
          "address": {
            "type": "string",
            "example": "Россия, Москва, Парусный проезд, 11"
          },
          "no_payments": {
            "type": "boolean",
            "default": false
          },
          "active": {
            "type": "boolean",
            "default": true
          },
          "phone": {
            "type": "string",
            "maxLength": 200,
            "example": "Дочь: 79999888776, соседка: 79874743384 (Зоя Николаевна)"
          },
          "division_name": {
            "readOnly": true,
            "type": "string"
          }
        },
        "required": [
          "address",
          "division",
          "latitude",
          "longitude",
          "name"
        ]
      },
      "ActiveObject": {
        "type": "object",
        "properties": {
          "object_id": {
            "type": "integer",
            "readOnly": true
          },
          "address": {
            "type": "string"
          },
          "division": {
            "type": "integer"
          },
          "division_name": {
            "readOnly": true,
            "type": "string"
          },
          "name": {
            "type": "string",
            "maxLength": 100
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
}